windows: provide an API for associating an OVERLAPPED file handle with the libuv event loop so that ReadFile and WriteFile can be called manually in the main event thread.
This makes possible support for OpenVPN TAP-Windows driver which
requires ReadFile calls with a large enough buffer to take an entire
packet and thus cannot be treated as a pipe as can be done on other
operating systems.

Since there are likely other "weird" files which need other kinds of
special treatment, this patch leaves the actual ReadFile() and
WriteFile() calls up to the user, providing a similar API to uv_poll.

Conflicts:
	AUTHORS
	include/uv-unix.h
	src/unix/core.c
	src/win/req-inl.h
	test/test-list.h